dev: extract retrier
