+++ pkg/util/runtime/runtime.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package util
+package runtime
 
 import (
 	"path"
